backend/requirements.txt
fastapi==0.112.2
uvicorn==0.30.6
sqlalchemy==2.0.36
pydantic==2.8.2
pydantic-settings==2.5.2
passlib[bcrypt]==1.7.4
python-jose==3.3.0
httpx==0.27.2
pytest==8.3.2
APScheduler==3.10.4
